Move error_code to ErrorCodes.h
