+++ rpcs3/Emu/Cell/Modules/cellMsgDialog.h
@@ -2,6 +2,7 @@
 
 #include "Utilities/BitField.h"
 #include "Emu/Memory/vm_ptr.h"
+#include "Emu/Cell/ErrorCodes.h"
 
 enum
 {
