+++ rpcs3/Emu/Cell/Modules/cellRtc.h
@@ -2,6 +2,7 @@
 
 #include "Utilities/BEType.h"
 #include "Emu/Memory/vm_ptr.h"
+#include "Emu/Cell/ErrorCodes.h"
 
 // Return Codes
 enum CellRtcError
