+++ rpcs3/Emu/Cell/PPUThread.h
@@ -318,20 +318,6 @@ struct ppu_gpr_cast_impl<b8, void>
 	}
 };
 
-template<>
-struct ppu_gpr_cast_impl<error_code, void>
-{
-	static inline u64 to(const error_code& code)
-	{
-		return code;
-	}
-
-	static inline error_code from(const u64 reg)
-	{
-		return not_an_error(reg);
-	}
-};
-
 template<typename T, typename AT>
 struct ppu_gpr_cast_impl<vm::_ptr_base<T, AT>, void>
 {
