+++ rpcs3/Emu/Cell/lv2/sys_memory.h
@@ -2,6 +2,7 @@
 
 #include "Emu/Memory/vm_ptr.h"
 #include "Emu/IdManager.h"
+#include "Emu/Cell/ErrorCodes.h"
 
 class cpu_thread;
