+++ rpcs3/Emu/Cell/lv2/sys_rsx.h
@@ -2,6 +2,7 @@
 
 #include "Utilities/mutex.h"
 #include "Emu/Memory/vm_ptr.h"
+#include "Emu/Cell/ErrorCodes.h"
 
 class cpu_thread;
