+++ rpcs3/Emu/Cell/lv2/sys_spu.h
@@ -3,6 +3,7 @@
 #include "sys_sync.h"
 #include "sys_event.h"
 #include "Emu/Cell/SPUThread.h"
+#include "Emu/Cell/ErrorCodes.h"
 
 #include "Emu/Memory/vm_ptr.h"
 #include "Utilities/File.h"
