+++ rpcs3/Emu/Cell/lv2/sys_storage.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #include "Emu/Memory/vm_ptr.h"
+#include "Emu/Cell/ErrorCodes.h"
 
 struct StorageDeviceInfo
 {
